+++ src/core/file_sys/quasifs/quasifs_inode.h
@@ -101,7 +101,7 @@ public:
         return -POSIX_EINVAL;
     }
 
-    virtual s64 getdents(void* buf, u32 count, s64 offset, s64* basep) {
+    virtual s64 getdents(void* buf, u64 count, s64 offset, s64* basep) {
         return -POSIX_EINVAL;
     }
