+++ src/core/file_sys/quasifs/quasifs_inode_quasi_directory_pfs.h
@@ -56,7 +56,7 @@ public:
     s64 pread(void* buf, u64 count, s64 offset) final override;
     s64 lseek(s64 current, s64 offset, s32 whence) final override;
 
-    s64 getdents(void* buf, u32 count, s64 offset, s64* basep) final override;
+    s64 getdents(void* buf, u64 count, s64 offset, s64* basep) final override;
 
     // Find an element with [name, case insensitive]
     inode_ptr lookup(const std::string& name) final override;
